requirements.txt
pandas
numpy
langchain
langchain-community
langchain-groq
sentence-transformers
chromadb
pymupdf
python-dotenv
streamlit